+++ pkg/backend/httpstate/backend.go
@@ -26,6 +26,7 @@ import (
 	"net/url"
 	"os"
 	"path"
+	"regexp"
 	"runtime"
 	"strconv"
 	"strings"
@@ -143,6 +144,7 @@ type cloudBackend struct {
 	url             string
 	stackConfigFile string
 	client          *client.Client
+	currentProject  *workspace.Project
 }
 
 // New creates a new Pulumi backend for the given cloud API URL and token.
@@ -153,11 +155,18 @@ func New(d diag.Sink, cloudURL, stackConfigFile string) (Backend, error) {
 		return nil, errors.Wrap(err, "getting stored credentials")
 	}
 
+	// When stringifying backend references, we take the current project (if present) into account.
+	currentProject, err := workspace.DetectProject()
+	if err != nil {
+		currentProject = nil
+	}
+
 	return &cloudBackend{
 		d:               d,
 		url:             cloudURL,
 		stackConfigFile: stackConfigFile,
 		client:          client.NewClient(cloudURL, apiToken, d),
+		currentProject:  currentProject,
 	}, nil
 }
 
@@ -378,6 +387,7 @@ func (b *cloudBackend) CloudURL() string { return b.url }
 func (b *cloudBackend) ParseStackReference(s string) (backend.StackReference, error) {
 	split := strings.Split(s, "/")
 	var owner string
+	var projectName string
 	var stackName string
 
 	if len(split) == 1 {
@@ -385,6 +395,10 @@ func (b *cloudBackend) ParseStackReference(s string) (backend.StackReference, er
 	} else if len(split) == 2 {
 		owner = split[0]
 		stackName = split[1]
+	} else if len(split) == 3 {
+		owner = split[0]
+		projectName = split[1]
+		stackName = split[2]
 	} else {
 		return nil, errors.Errorf("could not parse stack name '%s'", s)
 	}
@@ -397,10 +411,20 @@ func (b *cloudBackend) ParseStackReference(s string) (backend.StackReference, er
 		owner = currentUser
 	}
 
+	if projectName == "" {
+		currentProject, projectErr := workspace.DetectProject()
+		if projectErr != nil {
+			return nil, projectErr
+		}
+
+		projectName = currentProject.Name.String()
+	}
+
 	return cloudBackendReference{
-		owner: owner,
-		name:  tokens.QName(stackName),
-		b:     b,
+		owner:   owner,
+		project: projectName,
+		name:    tokens.QName(stackName),
+		b:       b,
 	}, nil
 }
 
@@ -433,7 +457,7 @@ func serveBrowserLoginServer(l net.Listener, expectedNonce string, destinationUR
 // CloudConsoleStackPath returns the stack path components for getting to a stack in the cloud console.  This path
 // must, of course, be combined with the actual console base URL by way of the CloudConsoleURL function above.
 func (b *cloudBackend) cloudConsoleStackPath(stackID client.StackIdentifier) string {
-	return path.Join(stackID.Owner, stackID.Stack)
+	return path.Join(stackID.Owner, stackID.Project, stackID.Stack)
 }
 
 // Logout logs out of the target cloud URL.
@@ -979,6 +1003,17 @@ func (b *cloudBackend) ImportDeployment(ctx context.Context, stackRef backend.St
 	return nil
 }
 
+var (
+	projectNameCleanRegexp = regexp.MustCompile("[^a-zA-Z0-9-_.]")
+)
+
+// cleanProjectName replaces undesirable characters in project names with hypens. At some point, these restrictions
+// will be further enfornced by the service, but for now we need to ensure that if we are making a rest call, we
+// do this cleaning on our end.
+func cleanProjectName(projectName string) string {
+	return projectNameCleanRegexp.ReplaceAllString(projectName, "-")
+}
+
 // getCloudStackIdentifier converts a backend.StackReference to a client.StackIdentifier for the same logical stack
 func (b *cloudBackend) getCloudStackIdentifier(stackRef backend.StackReference) (client.StackIdentifier, error) {
 	cloudBackendStackRef, ok := stackRef.(cloudBackendReference)
@@ -987,8 +1022,9 @@ func (b *cloudBackend) getCloudStackIdentifier(stackRef backend.StackReference) 
 	}
 
 	return client.StackIdentifier{
-		Owner: cloudBackendStackRef.owner,
-		Stack: string(cloudBackendStackRef.name),
+		Owner:   cloudBackendStackRef.owner,
+		Project: cleanProjectName(cloudBackendStackRef.project),
+		Stack:   string(cloudBackendStackRef.name),
 	}, nil
 }