+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Improvements
 
+- Add `--JSON` to `pulumi config` and `pulumi config get` to request the output be in JSON.
+
 ## 0.16.11 (Released January 16th, 2019)
 
 ### Improvements